+++ src/Rarus/BonusServer/Transactions/DTO/Operations/OperationCollection.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace Rarus\BonusServer\Transactions\DTO\Operations;
+
+/**
+ * Class OperationCollection
+ *
+ *
+ * @method  attach(Operation $operation, $data = null)
+ * @method  Operation current()
+ *
+ * @package Rarus\BonusServer\Transactions\DTO\Operations
+ */
+class OperationCollection extends \SplObjectStorage
+{
+}