+++ src/Rarus/BonusServer/Transactions/DTO/Type/Fabric.php
@@ -3,6 +3,8 @@ declare(strict_types=1);
 
 namespace Rarus\BonusServer\Transactions\DTO\Type;
 
+use Rarus\BonusServer\Exceptions\ApiClientException;
+
 /**
  * Class Fabric
  *
@@ -10,6 +12,27 @@ namespace Rarus\BonusServer\Transactions\DTO\Type;
  */
 class Fabric
 {
+    /**
+     * @param string $operationCode
+     *
+     * @return Type
+     * @throws ApiClientException
+     */
+    public static function initFromServerResponse(string $operationCode): Type
+    {
+        switch (strtolower($operationCode)) {
+            case 'sale':
+                return self::getSale();
+                break;
+            case 'refund':
+                return self::getRefund();
+                break;
+            default:
+                throw  new ApiClientException(sprintf('unknown operation code [%s]', $operationCode));
+                break;
+        }
+    }
+
     /**
      * @return Type
      */
@@ -21,7 +44,7 @@ class Fabric
     /**
      * @return Type
      */
-    public function getRefund(): Type
+    public static function getRefund(): Type
     {
         return new Type('refund');
     }